# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Hôtelière",
        "description": "Documentation de l'API du système de gestion hôtelière",
        "contact": {
            "name": "Support API",
            "email": "contact@hotel.com"
        },
        "license": {
            "name": "Licence propriétaire",
            "url": "https://www.hotel.com/licence"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Environnement de l'API"
        }
    ],
    "paths": {
...
        "/api/v1/admin/messages/inbox": {
            "get": {
                "tags": [
                    "Messages (Admin)"
                ],
                "summary": "Récupérer les messages reçus",
                "description": "Récupère la liste des messages reçus avec possibilité de filtrage",
                "operationId": "getInboxMessages",
                "parameters": [
                    {
                        "name": "read",
                        "in": "query",
                        "description": "Filtrer par statut de lecture",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sender_id",
                        "in": "query",
                        "description": "Filtrer par expéditeur",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans le sujet ou le contenu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des messages reçus récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Message"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès interdit - Droits administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/messages/sent": {
            "get": {
                "tags": [
                    "Messages (Admin)"
                ],
                "summary": "Récupérer les messages envoyés",
                "description": "Récupère la liste des messages envoyés avec possibilité de filtrage",
                "operationId": "getSentMessages",
                "parameters": [
                    {
                        "name": "recipient_id",
                        "in": "query",
                        "description": "Filtrer par destinataire",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans le sujet ou le contenu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des messages envoyés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Message"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès interdit - Droits administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "title": "Message",
                "description": "Modèle de message interne",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sender_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "recipient_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "subject": {
                        "type": "string",
                        "example": "Question sur ma réservation"
                    },
                    "content": {
                        "type": "string",
                        "example": "Bonjour, j'aimerais savoir s'il est possible de modifier les dates de ma réservation..."
                    },
                    "read": {
                        "type": "boolean",
                        "example": false
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "deleted_by_sender": {
                        "type": "boolean",
                        "example": false
                    },
                    "deleted_by_recipient": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "sender": {
                        "$ref": "#/components/schemas/User"
                    },
                    "recipient": {
                        "$ref": "#/components/schemas/User"
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Message"
                            }
                        ],
                        "nullable": true
                    },
                    "replies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez un token JWT pour s'authentifier",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}